WaterLily

Introduction and Quickstart

See the WaterLily README for an Introduction and Quickstart.

Types Methods and Functions

WaterLily.AbstractBody

WaterLily.AbstractPoisson

WaterLily.AutoBody

WaterLily.Bodies

WaterLily.Flow

WaterLily.MultiLevelPoisson

WaterLily.NoBody

WaterLily.Simulation

WaterLily.BC!

WaterLily.BC!

WaterLily.BCTuple

WaterLily.CIj

WaterLily.Jacobi!

WaterLily.L₂

WaterLily.accelerate!

WaterLily.apply!

WaterLily.curl

WaterLily.curvature

WaterLily.exitBC!

WaterLily.inside

WaterLily.inside_u

WaterLily.ke

WaterLily.loc

WaterLily.measure

WaterLily.measure!

WaterLily.measure!

WaterLily.measure_sdf!

WaterLily.mom_step!

WaterLily.mult!

WaterLily.pcg!

WaterLily.reduce_sdf_map

WaterLily.residual!

WaterLily.sdf

WaterLily.sdf

WaterLily.sdf_map_d

WaterLily.sim_step!

WaterLily.sim_time

WaterLily.slice

WaterLily.solver!

WaterLily.δ

WaterLily.λ₂

WaterLily.ω

WaterLily.ω_mag

WaterLily.ω_θ

WaterLily.∂

WaterLily.∮nds

WaterLily.@inside

WaterLily.@loop

WaterLily.AbstractBody — Type

AbstractBody

Immersed body Abstract Type. Any AbstractBody subtype must implement

d = sdf(body::AbstractBody, x, t=0)

and

d,n,V = measure(body::AbstractBody, x, t=0)

where d is the signed distance from x to the body at time t, and n & V are the normal and velocity vectors implied at x.

WaterLily.AbstractPoisson — Type

Poisson{N,M}

Composite type for conservative variable coefficient Poisson equations:

∮ds β ∂x/∂n = σ

The resulting linear system is

Ax = [L+D+L']x = z

where A is symmetric, block-tridiagonal and extremely sparse. Moreover, D[I]=-∑ᵢ(L[I,i]+L'[I,i]). This means matrix storage, multiplication, ect can be easily implemented and optimized without external libraries.

To help iteratively solve the system above, the Poisson structure holds helper arrays for inv(D), the error ϵ, and residual r=z-Ax. An iterative solution method then estimates the error ϵ=̃A⁻¹r and increments x+=ϵ, r-=Aϵ.

WaterLily.AutoBody — Type

AutoBody(sdf,map=(x,t)->x; compose=true) <: AbstractBody

sdf(x::AbstractVector,t::Real)::Real: signed distance function

map(x::AbstractVector,t::Real)::AbstractVector: coordinate mapping function

compose::Bool=true: Flag for composing sdf=sdf∘map

Implicitly define a geometry by its sdf and optional coordinate map. Note: the map is composed automatically if compose=true, i.e. sdf(x,t) = sdf(map(x,t),t). Both parameters remain independent otherwise. It can be particularly heplful to set compose=false when adding mulitple bodies together to create a more complex one.

WaterLily.Bodies — Type

Bodies(bodies, ops::AbstractVector)

bodies::Vector{AutoBody}: Vector of AutoBody

ops::Vector{Function}: Vector of operators for the superposition of multiple AutoBodys

Superposes multiple body::AutoBody objects together according to the operators ops. While this can be manually performed by the operators implemented for AutoBody, adding too many bodies can yield a recursion problem of the sdf and map functions not fitting in the stack. This type implements the superposition of bodies by iteration instead of recursion, and the reduction of the sdf and map functions is done on the mesure function, and not before. The operators vector ops specifies the operation to call between two consecutive AutoBodys in the bodies vector. Note that + (or the alias ∪) is the only operation supported between Bodies.

WaterLily.Flow — Type

Flow{D::Int, T::Float, Sf<:AbstractArray{T,D}, Vf<:AbstractArray{T,D+1}, Tf<:AbstractArray{T,D+2}}

Composite type for a multidimensional immersed boundary flow simulation.

Flow solves the unsteady incompressible Navier-Stokes equations on a Cartesian grid. Solid boundaries are modelled using the Boundary Data Immersion Method. The primary variables are the scalar pressure p (an array of dimension D) and the velocity vector field u (an array of dimension D+1).

WaterLily.MultiLevelPoisson — Type

MultiLevelPoisson{N,M}

Composite type used to solve the pressure Poisson equation with a geometric multigrid method. The only variable is levels, a vector of nested Poisson systems.

WaterLily.NoBody — Type

NoBody

Use for a simulation without a body.

WaterLily.Simulation — Type

Simulation(dims::NTuple, u_BC::Union{NTuple,Function}, L::Number;
           U=norm2(u_BC), Δt=0.25, ν=0., ϵ=1, perdir=(1,)
           uλ::nothing, g=nothing, exitBC=false,
           body::AbstractBody=NoBody(),
           T=Float32, mem=Array)

Constructor for a WaterLily.jl simulation:

dims: Simulation domain dimensions.

u_BC: Simulation domain velocity boundary conditions, either a tuple u_BC[i]=uᵢ, i=eachindex(dims), or a time-varying function f(i,t)

L: Simulation length scale.

U: Simulation velocity scale.

Δt: Initial time step.

ν: Scaled viscosity (Re=UL/ν).

g: Domain acceleration, g(i,t)=duᵢ/dt

ϵ: BDIM kernel width.

perdir: Domain periodic boundary condition in the (i,) direction.

exitBC: Convective exit boundary condition in the i=1 direction.

uλ: Function to generate the initial velocity field.

body: Immersed geometry.

T: Array element type.

mem: memory location. Array, CuArray, ROCm to run on CPU, NVIDIA, or AMD devices, respectively.

See files in examples folder for examples.

WaterLily.BC! — Function

BC!(a,A)

Apply boundary conditions to the ghost cells of a vector field. A Dirichlet condition a[I,i]=A[i] is applied to the vector component normal to the domain boundary. For example aₓ(x)=Aₓ ∀ x ∈ minmax(X). A zero Neumann condition is applied to the tangential components.

WaterLily.BC! — Method

BC!(a)

Apply zero Neumann boundary conditions to the ghost cells of a scalar field.

WaterLily.BCTuple — Method

BCTuple(f,t,N)

Generate a tuple of N values from either a boundary condition function f(i,t) or the tuple of boundary conditions f=(fₓ,...).

WaterLily.CIj — Method

CIj(j,I,jj)

Replace jᵗʰ component of CartesianIndex with k

WaterLily.Jacobi! — Method

Jacobi!(p::Poisson; it=1)

Jacobi smoother run it times. Note: This runs for general backends, but is very slow to converge.

WaterLily.L₂ — Method

L₂(a)

L₂ norm of array a excluding ghosts.

WaterLily.accelerate! — Method

accelerate!(r,t,g)

This function adds a uniform acceleration field g at time t to r. If g ≠ nothing, then g(i,t)=dUᵢ/dt.

WaterLily.apply! — Method

apply!(f, c)

Apply a vector function f(i,x) to the faces of a uniform staggered array c or a function f(x) to the center of a uniform array c.

WaterLily.curl — Method

curl(i,I,u)

Compute component i of $𝛁×𝐮$ at the edge of cell I. For example curl(3,CartesianIndex(2,2,2),u) will compute ω₃(x=1.5,y=1.5,z=2) as this edge produces the highest accuracy for this mix of cross derivatives on a staggered grid.

WaterLily.curvature — Method

curvature(A::AbstractMatrix)

Return H,K the mean and Gaussian curvature from A=hessian(sdf). K=tr(minor(A)) in 3D and K=0 in 2D.

WaterLily.exitBC! — Method

exitBC!(u,u⁰,U,Δt)

Apply a 1D convection scheme to fill the ghost cell on the exit of the domain.

WaterLily.inside — Method

inside(a)

Return CartesianIndices range excluding a single layer of cells on all boundaries.

WaterLily.inside_u — Method

inside_u(dims,j)

Return CartesianIndices range excluding the ghost-cells on the boundaries of a vector array on face j with size dims.

WaterLily.ke — Method

ke(I::CartesianIndex,u,U=0)

Compute $½∥𝐮-𝐔∥²$ at center of cell I where U can be used to subtract a background flow (by default, U=0).

WaterLily.loc — Method

loc(i,I) = loc(Ii)

Location in space of the cell at CartesianIndex I at face i. Using i=0 returns the cell center s.t. loc = I.

WaterLily.measure! — Function

measure!(sim::Simulation,t=timeNext(sim))

Measure a dynamic body to update the flow and pois coefficients.

WaterLily.measure! — Method

measure!(flow::Flow, body::AbstractBody; t=0, ϵ=1)

Queries the body geometry to fill the arrays:

flow.μ₀, Zeroth kernel moment

flow.μ₁, First kernel moment scaled by the body normal

flow.V, Body velocity

at time t using an immersion kernel of size ϵ.

See Maertens & Weymouth, doi:10.1016/j.cma.2014.09.007.

WaterLily.measure — Method

d,n,V = measure(body::AutoBody,x,t)
d,n,V = measure(body::Bodies,x,t)

Determine the implicit geometric properties from the sdf and map. The gradient of d=sdf(map(x,t)) is used to improve d for pseudo-sdfs. The velocity is determined solely from the optional map function.

WaterLily.measure_sdf! — Function

measure_sdf!(a::AbstractArray, body::AbstractBody, t=0)

Uses sdf(body,x,t) to fill a.

WaterLily.mom_step! — Method

mom_step!(a::Flow,b::AbstractPoisson)

Integrate the Flow one time step using the Boundary Data Immersion Method and the AbstractPoisson pressure solver to project the velocity onto an incompressible flow.

WaterLily.mult! — Method

mult!(p::Poisson,x)

Efficient function for Poisson matrix-vector multiplication. Fills p.z = p.A x with 0 in the ghost cells.

WaterLily.pcg! — Method

pcg!(p::Poisson; it=6)

Conjugate-Gradient smoother with Jacobi preditioning. Runs at most it iterations, but will exit early if the Gram-Schmidt update parameter |α| < 1% or |r D⁻¹ r| < 1e-8. Note: This runs for general backends and is the default smoother.

WaterLily.reduce_sdf_map — Method

reduce_sdf_map(sdf_a,map_a,d_a,sdf_b,map_b,d_b,op,x,t)

Reduces two different sdf and map functions, and d value.

WaterLily.residual! — Method

residual!(p::Poisson)

Computes the resiual r = z-Ax and corrects it such that r = 0 if iD==0 which ensures local satisfiability and sum(r) = 0 which ensures global satisfiability.

The global correction is done by adjusting all points uniformly, minimizing the local effect. Other approaches are possible.

Note: These corrections mean x is not strictly solving Ax=z, but without the corrections, no solution exists.

WaterLily.sdf — Method

d = sdf(body::AutoBody,x,t) = body.sdf(x,t)

WaterLily.sdf — Method

d = sdf(a::Bodies,x,t)

Computes distance for Bodies type.

WaterLily.sdf_map_d — Method

sdf_map_d(ab::Bodies,x,t)

Returns the sdf and map functions, and the distance d (d=sdf(x,t)) for the Bodies type.

WaterLily.sim_step! — Method

sim_step!(sim::Simulation,t_end=sim(time)+Δt;max_steps=typemax(Int),remeasure=true,verbose=false)

Integrate the simulation sim up to dimensionless time t_end. If remeasure=true, the body is remeasured at every time step. Can be set to false for static geometries to speed up simulation.

WaterLily.sim_time — Method

sim_time(sim::Simulation)

Return the current dimensionless time of the simulation tU/L where t=sum(Δt), and U,L are the simulation velocity and length scales.

WaterLily.slice — Method

slice(dims,i,j,low=1)

Return CartesianIndices range slicing through an array of size dims in dimension j at index i. low optionally sets the lower extent of the range in the other dimensions.

WaterLily.solver! — Method

solver!(A::Poisson;log,tol,itmx)

Approximate iterative solver for the Poisson matrix equation Ax=b.

A: Poisson matrix with working arrays.

A.x: Solution vector. Can start with an initial guess.

A.z: Right-Hand-Side vector. Will be overwritten!

A.n[end]: stores the number of iterations performed.

log: If true, this function returns a vector holding the L₂-norm of the residual at each iteration.

tol: Convergence tolerance on the L₂-norm residual.

itmx: Maximum number of iterations.

WaterLily.δ — Method

δ(i,N::Int)
δ(i,I::CartesianIndex{N}) where {N}

Return a CartesianIndex of dimension N which is one at index i and zero elsewhere.

WaterLily.λ₂ — Method

λ₂(I::CartesianIndex{3},u)

λ₂ is a deformation tensor metric to identify vortex cores. See https://en.wikipedia.org/wiki/Lambda2_method and Jeong, J., & Hussain, F., doi:10.1017/S0022112095000462

WaterLily.ω — Method

ω(I::CartesianIndex{3},u)

Compute 3-vector $𝛚=𝛁×𝐮$ at the center of cell I.

WaterLily.ω_mag — Method

ω_mag(I::CartesianIndex{3},u)

Compute $∥𝛚∥$ at the center of cell I.

WaterLily.ω_θ — Method

ω_θ(I::CartesianIndex{3},z,center,u)

Compute $𝛚⋅𝛉$ at the center of cell I where $𝛉$ is the azimuth direction around vector z passing through center.

WaterLily.∂ — Method

∂(i,j,I,u)

Compute $∂uᵢ/∂xⱼ$ at center of cell I. Cross terms are computed less accurately than inline terms because of the staggered grid.

WaterLily.∮nds — Method

∮nds(p,body::AutoBody,t=0)

Surface normal integral of field p over the body.

WaterLily.@inside — Macro

@inside <expr>

Simple macro to automate efficient loops over cells excluding ghosts. For example,

@inside p[I] = sum(loc(0,I))

becomes

@loop p[I] = sum(loc(0,I)) over I ∈ inside(p)

See @loop.

WaterLily.@loop — Macro

@loop <expr> over <I ∈ R>

Macro to automate fast CPU or GPU loops using KernelAbstractions.jl. The macro creates a kernel function from the expression <expr> and evaluates that function over the CartesianIndices I ∈ R.

For example

@loop a[I,i] += sum(loc(i,I)) over I ∈ R

becomes

@kernel function kern(a,i,@Const(I0))
    I ∈ @index(Global,Cartesian)+I0
    a[I,i] += sum(loc(i,I))
end
kern(get_backend(a),64)(a,i,R[1]-oneunit(R[1]),ndrange=size(R))

where get_backend is used on the first variable in expr (a in this example).