WaterLily

Introduction and Quickstart

See the WaterLily README for an Introduction and Quickstart.

Types Methods and Functions

WaterLily.AbstractBody

WaterLily.AbstractPoisson

WaterLily.AutoBody

WaterLily.Flow

WaterLily.MultiLevelPoisson

WaterLily.NoBody

WaterLily.Simulation

WaterLily.BC!

WaterLily.BC!

WaterLily.Jacobi!

WaterLily.L₂

WaterLily.apply!

WaterLily.curl

WaterLily.curvature

WaterLily.inside

WaterLily.inside_u

WaterLily.ke

WaterLily.loc

WaterLily.measure

WaterLily.measure!

WaterLily.measure!

WaterLily.measure_sdf!

WaterLily.mom_step!

WaterLily.mult!

WaterLily.pcg!

WaterLily.sim_step!

WaterLily.sim_time

WaterLily.slice

WaterLily.solver!

WaterLily.δ

WaterLily.λ₂

WaterLily.ω

WaterLily.ω_mag

WaterLily.ω_θ

WaterLily.∂

WaterLily.∮nds

WaterLily.@inside

WaterLily.@loop

WaterLily.AbstractBody — Type

AbstractBody

Immersed body Abstract Type. Any AbstractBody subtype must implement

`measure!(flow::Flow, body::AbstractBody; t=0, ϵ=1)`

which queries the body geometry to fill the arrays:

`flow.μ₀`, Zeroth kernel moment
`flow.μ₁`, First kernel moment scaled by the body normal
`flow.V`,  Body velocity
`flow.σᵥ`, Body velocity divergence scaled by `μ₀-1`

at time t using an immersion kernel of size ϵ. See Maertens & Weymouth, https://doi.org/10.1016/j.cma.2014.09.007

WaterLily.AbstractPoisson — Type

Poisson{N,M}

Composite type for conservative variable coefficient Poisson equations:

∮ds β ∂x/∂n = σ

The resulting linear system is

Ax = [L+D+L']x = z

where A is symmetric, block-tridiagonal and extremely sparse. Moreover, D[I]=-∑ᵢ(L[I,i]+L'[I,i]). This means matrix storage, multiplication, ect can be easily implemented and optimized without external libraries.

To help iteratively solve the system above, the Poisson structure holds helper arrays for inv(D), the error ϵ, and residual r=z-Ax. An iterative solution method then estimates the error ϵ=̃A⁻¹r and increments x+=ϵ, r-=Aϵ.

WaterLily.AutoBody — Type

AutoBody(sdf,map=(x,t)->x; compose=true) <: AbstractBody

- sdf(x::AbstractVector,t::Real)::Real: signed distance function
- map(x::AbstractVector,t::Real)::AbstractVector: coordinate mapping function
- compose::Bool: if true, automatically compose the `map`, ie `sdf(x,t) = sdf(map(x,t),t)`
                    else, `sdf` and `map` remain independent

Define a geometry by its sdf and optional coordinate map. All other properties are determined using Automatic Differentiation. Note: the map is composed automatically if compose is set to true, ie sdf(x,t) = sdf(map(x,t),t). Both parameters remain independent otherwise. It can be particularly heplful to set it as false when adding mulitple bodies together to create a more complexe one.

WaterLily.Flow — Type

Flow{D, V, S, F, B, T}

Composite type for a multidimensional immersed boundary flow simulation.

Flow solves the unsteady incompressible Navier-Stokes equations on a Cartesian grid. Solid boundaries are modelled using the Boundary Data Immersion Method. The primary variables are the scalar pressure p (an array of dimension N) and the velocity vector field u (an array of dimension M=N+1).

WaterLily.MultiLevelPoisson — Type

MultiLevelPoisson{N,M}

Composite type used to solve the pressure Poisson equation with a geometric multigrid method. The only variable is levels, a vector of nested Poisson systems.

WaterLily.Simulation — Type

Simulation(dims::NTuple, u_BC::NTuple, L::Number;
           U=norm2(u_BC), Δt=0.25, ν=0., ϵ = 1,
           uλ::Function=(i,x)->u_BC[i],
           body::AbstractBody=NoBody(),
           T=Float32, mem = Array)

Constructor for a WaterLily.jl simulation:

`dims`: Simulation domain dimensions.
`u_BC`: Simulation domain velocity boundary conditions, `u_BC[i]=uᵢ, i=eachindex(dims)`.
`L`: Simulation length scale.
`U`: Simulation velocity scale.
`Δt`: Initial time step.
`ν`: Scaled viscosity (`Re=UL/ν`).
`ϵ`: BDIM kernel width.
`uλ`: Function to generate the initial velocity field.
`body`: Immersed geometry.
`T`: Array element type.
`mem`: memory location. `Array` and `CuArray` run on CPU and CUDA backends, respectively.

See files in examples folder for examples.

WaterLily.BC! — Function

BC!(a,A,f=1)

Apply boundary conditions to the ghost cells of a vector field. A Dirichlet condition a[I,i]=f*A[i] is applied to the vector component normal to the domain boundary. For example aₓ(x)=f*Aₓ ∀ x ∈ minmax(X). A zero Nuemann condition is applied to the tangential components.

WaterLily.BC! — Method

BC!(a)

Apply zero Nuemann boundary conditions to the ghost cells of a scalar field.

WaterLily.L₂ — Method

L₂(a)

L₂ norm of array a excluding ghosts.

WaterLily.apply! — Method

apply!(f, c)

Apply a vector function f(i,x) to the faces of a uniform staggered array c.

WaterLily.inside — Method

inside(a)

Return CartesianIndices range excluding a single layer of cells on all boundaries.

WaterLily.loc — Method

loc(i,I)

Location in space of the cell at CartesianIndex I at face i. Using i=0 returns the cell center s.t. loc = I.

WaterLily.measure! — Function

measure!(sim::Simulation,t=time(sim))

Measure a dynamic body to update the flow and pois coefficients.

WaterLily.measure! — Method

measure!(flow::Flow, body::AutoBody; t=0, ϵ=1)

Uses body.sdf and body.map to fill the arrays:

`flow.μ₀`, Zeroth kernel moment
`flow.μ₁`, First kernel moment scaled by the body normal
`flow.V`,  Body velocity
`flow.σᵥ`,  Body velocity divergence scaled by `μ₀-1`

at time t using an immersion kernel of size ϵ. See Maertens & Weymouth

WaterLily.measure — Method

measure(body::AutoBody,x,t)

ForwardDiff is used to determine the geometric properties from the sdf. Note: The velocity is determined soley from the optional map function.

WaterLily.mom_step! — Method

mom_step!(a::Flow,b::AbstractPoisson)

Integrate the Flow one time step using the Boundary Data Immersion Method and the AbstractPoisson pressure solver to project the velocity onto an incompressible flow.

WaterLily.mult! — Method

mult!(p::Poisson,x)

Efficient function for Poisson matrix-vector multiplication. Fills p.z = p.A x with 0 in the ghost cells.

WaterLily.sim_step! — Method

sim_step!(sim::Simulation,t_end;remeasure=true,verbose=false)

Integrate the simulation sim up to dimensionless time t_end. If remeasure=true, the body is remeasured at every time step. Can be set to false for static geometries to speed up simulation.

WaterLily.sim_time — Method

sim_time(sim::Simulation)

Return the current dimensionless time of the simulation tU/L where t=sum(Δt), and U,L are the simulation velocity and length scales.

WaterLily.solver! — Method

solver!(A::Poisson;log,tol,itmx)

Approximate iterative solver for the Poisson matrix equation Ax=b.

`A`: Poisson matrix with working arrays
`A.x`: Solution vector. Can start with an initial guess.
`A.z`: Right-Hand-Side vector. Will be overwritten! 
`A.n[end]`: stores the number of iterations performed.
`log`: If `true`, this function returns a vector holding the `L₂`-norm of the residual at each iteration.
`tol`: Convergence tolerance on the `L₂`-norm residual.
'itmx': Maximum number of iterations

WaterLily.δ — Method

δ(i,N::Int)
δ(i,I::CartesianIndex{N}) where {N}

Return a CartesianIndex of dimension N which is one at index i and zero elsewhere.

WaterLily.@inside — Macro

@inside <expr>

Simple macro to automate efficient loops over cells excluding ghosts. For example

@inside p[I] = sum(loc(0,I))

becomes

@loop p[I] = sum(loc(0,I)) over I ∈ inside(p)

See @loop.

WaterLily.NoBody — Type

NoBody

Use for a simulation without a body

WaterLily.Jacobi! — Method

Jacobi!(p::Poisson; it=1)

Jacobi smoother run it times. Note: This runs for general backends, but is very slow to converge.

WaterLily.curl — Method

curl(i,I,u)

Compute component i of ∇×u at the edge of cell I. For example curl(3,CartesianIndex(2,2,2),u) will compute ω₃(x=1.5,y=1.5,z=2) as this edge produces the highest accuracy for this mix of cross derivatives on a staggered grid.

WaterLily.curvature — Method

curvature(A::AbstractMatrix)

Return H,K the mean and Gaussian curvature from A=hessian(sdf). K=tr(minor(A)) in 3D and K=0 in 2D.

WaterLily.inside_u — Method

inside_u(dims,j)

Return CartesianIndices range excluding the ghost-cells on the boundaries of a vector array on face j with size dims.

WaterLily.ke — Method

ke(I::CartesianIndex,u,U=0)

Compute ½|u-U|² at center of cell I where U can be used to subtract a background flow.

WaterLily.measure_sdf! — Method

measure_sdf!(a::AbstractArray, body::AutoBody, t=0)

Uses body.sdf(x,t) to fill a.

WaterLily.pcg! — Method

pcg!(p::Poisson; it=6)

Conjugate-Gradient smoother with Jacobi preditioning. Runs at most it iterations, but will exit early if the Gram-Smit update parameter |α|<1% or |rD⁻¹r|<1e-8. Note: This runs for general backends and is the default smoother.

WaterLily.slice — Method

slice(dims,i,j,low=1,trim=0)

Return CartesianIndices range slicing through an array of size dims in dimension j at index i. low optionally sets the lower extent of the range in the other dimensions.

WaterLily.λ₂ — Method

λ₂(I::CartesianIndex{3},u)

λ₂ is a deformation tensor metric to identify vortex cores. See https://en.wikipedia.org/wiki/Lambda2_method and Jeong, J., & Hussain, F. doi:10.1017/S0022112095000462

WaterLily.ω — Method

ω(I::CartesianIndex{3},u)

Compute 3-vector ω=∇×u at the center of cell I.

WaterLily.ω_mag — Method

ω_mag(I::CartesianIndex{3},u)

Compute |ω| at the center of cell I.

WaterLily.ω_θ — Method

ω_θ(I::CartesianIndex{3},z,center,u)

Compute ω⋅θ at the center of cell I where θ is the azimuth direction around vector z passing through center.

WaterLily.∂ — Method

∂(i,j,I,u)

Compute ∂uᵢ/∂xⱼ at center of cell I. Cross terms are computed less accurately than inline terms because of the staggered grid.

WaterLily.∮nds — Method

∮nds(p,body::AutoBody,t=0)

Surface normal integral of field p over the body.

WaterLily.@loop — Macro

@loop <expr> over <I ∈ R>

Macro to automate fast CPU or GPU loops using KernelAbstractions.jl. The macro creates a kernel function from the expression <expr> and evaluates that function over the CartesianIndices I ∈ R.

For example

@loop a[I,i] += sum(loc(i,I)) over I ∈ R

becomes

@kernel function kern(a,i,@Const(I0))
    I ∈ @index(Global,Cartesian)+I0
    a[I,i] += sum(loc(i,I))
end
kern(get_backend(a),64)(a,i,R[1]-oneunit(R[1]),ndrange=size(R))

where get_backend is used on the first variable in expr (a in this example).